% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fqr.R
\name{coef.fqr}
\alias{coef.fqr}
\alias{coefficients.fqr}
\title{Get coefficients from fast quantile regression}
\usage{
\method{coef}{fqr}(object, ...)

\method{coefficients}{fqr}(object, ...)
}
\arguments{
\item{object}{object to get coefficients from}

\item{...}{additional arguments, ignored for now}
}
\description{
Get coefficients from fast quantile regression
}
