% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fqr.R
\name{predict.fqr}
\alias{predict.fqr}
\title{Predict from fast quantile regression fit}
\usage{
\method{predict}{fqr}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{a model object to use for prediction}

\item{newdata}{a data.frame with the same columns as data used to fit model}

\item{...}{other parameters, ignored for now}
}
\description{
Predict from fast quantile regression fit
}
