% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fqr.R
\name{print.fqr}
\alias{print.fqr}
\title{Print fast quantile regression}
\usage{
\method{print}{fqr}(x, digits = getOption("digits"), ...)
}
\arguments{
\item{x}{fit to print}

\item{digits}{digits to use when printing}

\item{...}{other arguments to print}
}
\description{
Print fast quantile regression
}
