% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fqr.R
\name{resid.fqr}
\alias{resid.fqr}
\alias{residuals.fqr}
\title{Get model residuals}
\usage{
\method{resid}{fqr}(object, ...)

\method{residuals}{fqr}(object, ...)
}
\arguments{
\item{object}{object to get coefficients from}

\item{...}{additional arguments, ignored for now}
}
\description{
Get model residuals
}
